+++ mmodel/__init__.py
@@ -1,4 +1,3 @@
 from mmodel.graph import MGraph
 from mmodel.model import Model, PlainModel, H5Model
 from mmodel.loop import basic_loop
-from mmodel.draw import draw_graph