+++ mmodel/doc.py
@@ -1,0 +1,14 @@
+from pydoc import Helper
+
+
+class ModelHelper(Helper):
+
+    """Inherite helper function to modify"""
+
+    def help(self, request):
+        """Add mmodel specific string to help function"""
+        self.output.write(request.doc_long)
+        super().help(request)
+
+
+helper = ModelHelper()