+++ mmodel/draw.py
@@ -1,68 +1,85 @@
 import graphviz
+import networkx as nx
+from copy import deepcopy
+
+DEFAULT_SETTINGS = {
+    "graph_attr": {
+        "labelloc": "t",
+        "labeljust": "l",
+        "splines": "ortho",
+        "ordering": "out",
+    },
+    "node_attr": {"shape": "box"},
+}
 
 
-def draw_graph(G, name=None, show_detail=True, filename=None, **kwargs):
-    """Show model graph
-
-    The process create a graphviz graph and write networkx.graph
-    nodes and edges to it.
-
-    :param str title: title of the graph
-    :param bool show_detail: default to true, all nodes signature
-        and returns are shown.
-
-    Optional dot graph settings can be suplied using additional
-    keyword arguments - graph_attr, node_attr, and edge_attr.
-    See graphviz dot file documentation for detailed customization.
-    The original setting is completely replaced. Note the title
-    (graph label) should be included in "graph_attr".
-
+def update_settings(label):
+    """Update graphviz settings
+    
+    Creates a copy of the default dictionary
+    and update the graph label in graph attribute
     """
 
-    name = name or G.name
-    doc = G.graph.get('doc', '')
-    filename = filename or f"{name}.gv"
-    default_settings = {
-        "graph_attr": {
-            "labelloc": "t",
-            "splines": "ortho",
-            "ordering": "out",
-        },
-        "node_attr": {"shape": "box"},
-    }
+    # copy() is shallow, does not copy the nested dict
+    new_settings = deepcopy(DEFAULT_SETTINGS)
+    new_settings["graph_attr"].update({"label": label})
 
-    settings = {**default_settings, **kwargs}
-    settings["graph_attr"]["label"] = f"{name}\n{doc}"
+    return new_settings
 
-    dot_graph = graphviz.Digraph(name=name, filename=filename, **settings)
 
-    dot_subgraphs = []
+def draw_plain_graph(G, name, label):
+    """Draw plain graph
+    
+    :param str name: name of the graph
+    :param str label: title of the graph
+
+    Plain graph contains the graph label (name + doc)
+    Each node only shows the node name
+    """
+
+    settings = update_settings(label)
+    dot_graph = graphviz.Digraph(name=name, **settings)
+
+    for node in G.nodes:
+        dot_graph.node(node)
+
+    for u, v in G.edges:
+        dot_graph.edge(u, v)
+
+    return dot_graph
+
+
+def draw_graph(G, name, label):
+    """Draw detailed graph
+    
+    :param str name: name of the graph
+    :param str label: title of the graph
+
+    Plain graph contains the graph label (name + doc + input + output)
+    Each node shows node label (name + signature + returns)
+
+    Subgraph node shows the label and subgraph doc.
+    """
+
+    settings = update_settings(label)
+    dot_graph = graphviz.Digraph(name=name, **settings)
 
     for node, ndict in G.nodes(data=True):
-        if ndict.get("has_subgraph", False):
-            dot_subgraphs.append([node, ndict])
-
-        if show_detail:
-            rts = ", ".join(ndict["return_params"])
-            label = f"{node}\l\n{ndict['node_obj'].__name__}{ndict['signature']}\lreturn {rts}\l"
-            dot_graph.node(node, label=label)
-        else:
-            dot_graph.node(node)
+        rts = ", ".join(ndict["return_params"])
+        label = f"{node}\l\n{ndict['node_obj'].__name__}{ndict['signature']}\lreturn {rts}\l"
+        dot_graph.node(node, label=label)
 
     for u, v, edict in G.edges(data=True):
-        if show_detail:
-            dot_graph.edge(u, v, xlabel=" ".join(edict["interm_params"]))
-        else:
-            dot_graph.edge(u, v)
+        dot_graph.edge(u, v, xlabel=" ".join(edict["interm_params"]))
 
-    # draw subgraph if there is any
-    for node, ndict in dot_subgraphs:
-        node_obj = ndict["node_obj"]
-        subgraph = node_obj.G
-        title = node
-        dot_sub = draw_graph(
-            subgraph, name=f"cluster_{title}", show_detail=show_detail, **kwargs
-        )
+    for node in nx.get_node_attributes(G, "has_subgraph").keys():
+
+        subG = G.nodes[node]["node_obj"].G
+
+        # use short docstring for subgraph
+        label = subG.doc_short.replace('\n', '\l')
+        dot_sub = draw_graph(subG, name=f"cluster {node}", label=label)
         dot_graph.subgraph(dot_sub)
 
     return dot_graph
+